requirements.txt
Flask==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
stripe==6.6.0
psycopg2
clize
bcrypt
flask_login
pytest
Flask-Mail
flask-sitemap
